+++ tedana/decomp/__init__.py
@@ -1,2 +1,11 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
+
+from .eigendecomp import (
+    tedpca, tedica, eimask,
+)
+
+
+__all__ = [
+    'tedpca', 'tedica', 'eimask'
+    ]
