+++ tedana/io/__init__.py
@@ -1,2 +1,19 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
+
+from .input_data import (
+    ctabsel,
+)
+
+
+from .output_data import (
+    gscontrol_mmix, split_ts, write_split_ts, writefeats,
+    writect, writeresults, writeresults_echoes,
+)
+
+
+__all__ = [
+    'ctabsel', 'gscontrol_mmix',
+    'split_ts', 'write_split_ts',
+    'writefeats', 'writect', 'writeresults',
+    'writeresults_echoes']
