+++ tedana/model/__init__.py
@@ -1,2 +1,22 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
+
+from .fit import (
+    computefeats2,
+    fitmodels_direct, get_coeffs,
+    getelbow_cons, getelbow_mod,
+    getelbow_aggr, gscontrol_raw,
+    spatclust,
+)
+
+
+from .t2smap import (
+    fit, make_optcom, t2sadmap,
+)
+
+
+__all__ = [
+    'computefeats2', 'fit', 'fitmodels_direct',
+    'get_coeffs', 'getelbow_cons', 'getelbow_mod',
+    'getelbow_aggr', 'gscontrol_raw',
+    'make_optcom', 'spatclust', 't2sadmap']
