+++ tedana/select/__init__.py
@@ -1,7 +1,9 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
 
-from .t2smap import (t2sadmap, make_optcom)
+from .select_comps import (
+    selcomps,
+)
 
-__all__ = [
-    't2sadmap', 'make_optcom']
+
+__all__ = ['selcomps']
