+++ tedana/workflows/__init__.py
@@ -1,2 +1,7 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
+
+from . import tedana, t2smap
+
+
+__all__ = ['tedana', 't2smap']
